+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.4
 * Add wp-env for local development.
 * Enforce WP linting across plugin codebase.
+* Use webpack generated dependencies.
 
 ## 1.0.3
 * Update README docs.